move interpolation into respective directories
