+++ code/tests/stf/CMakeLists.txt
@@ -3,7 +3,6 @@ enable_testing()
 set(TEST_FILES_PREFIX "cpp/stf")
 
 set(TEST_FILES
-    ${TEST_FILES_PREFIX}/alg/interpolation_tests.cpp
     ${TEST_FILES_PREFIX}/alg/intersection_tests.cpp
     ${TEST_FILES_PREFIX}/cam/scamera_tests.cpp
     ${TEST_FILES_PREFIX}/geom/aabb2_tests.cpp
@@ -11,6 +10,7 @@ set(TEST_FILES
     ${TEST_FILES_PREFIX}/gfx/color_tests.cpp
     ${TEST_FILES_PREFIX}/math/mtx4_tests.cpp
     ${TEST_FILES_PREFIX}/math/crange_tests.cpp
+    ${TEST_FILES_PREFIX}/math/interpolation_tests.cpp
     ${TEST_FILES_PREFIX}/math/range_tests.cpp
     ${TEST_FILES_PREFIX}/math/spherical_tests.cpp
     ${TEST_FILES_PREFIX}/math/transform_tests.cpp
